backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic[email]==2.9.2
apscheduler==3.10.4
